Update popup.js
I made a change so that for each script or media file the program catches, rather than linking to a WOT review for the domain that the file was hosted on, it provides direct download link for the file itself. This simplifies downloading mp4 files on many streaming sites. Before, one would have to allow the file, refresh, and then use a separate plugin to download the file. This feature speeds up the process, fully taking advantage of this plugin's ability to detect mp4 files.